+++ semana14/sistema-bancario/.gitignore
