+++ semana14/sistema-bancario/package.json
@@ -13,6 +13,10 @@
   "author": "Ricardo Araujo",
   "license": "ISC",
   "dependencies": {
+    "moment": "^2.26.0"
+  },
+  "devDependencies": {
+    "@types/moment": "^2.13.0",
     "@types/node": "^14.0.4"
   }
 }
