+++ semana14/sistema-bancario/src/errorMensages.ts
@@ -1,0 +1,3 @@
+export const errorMensage = "Não encontramos nenhuma conta correspondente";
+
+export const erroUnderage = "O usuario não tem mais de 18 anos";
