+++ semana14/sistema-bancario/src/index.ts
@@ -1,27 +1,12 @@
-import * as fs from "fs";
-
-const accountsFile = require("../data/accounts.json");
-
-type account = {
-    name: string,
-    cpf: string,
-    birthDate: string,
-    balance: number,
-    statement: statement[]
-}
-
-type statement = {
-    operation: string,
-    data: any
-}
-
+import {account} from "./types"
+import createAccount from "./services/createAccount";
+import addBalance from "./services/addBalance";
+import getBalance from "./services/getBalance";
+import transference from "./services/transference";
 
 const action: string = process.argv[2];
-const errorMensage = "Não encontramos nenhuma conta correspondente"
-
 
 switch(action){
-
     case "createAccount":{
         
         const mockAccount: account = {
@@ -32,89 +17,21 @@ switch(action){
             "statement": []
         }
 
-        if(mockAccount.name && mockAccount.cpf && mockAccount.birthDate && mockAccount.balance){
-            
-            const checkCPF = accountsFile.find((account)=>{
-                        return account.cpf === mockAccount.cpf
-                    })
-            
-            if(checkCPF===undefined){
-               
-                const neWFile = accountsFile;
-    
-                neWFile.push(mockAccount);
-        
-                try{
-                    fs.writeFileSync(`data/accounts.json`, JSON.stringify(neWFile));
-                    console.log('Criado com sucesso');
-                    break;
-                }catch(error){
-                    console.error(error);
-                    break;
-                }
-            }else{
-                console.log("Já existe uma conta com este associado cpf!");
-            }
-        }else{
-            console.log("digite um nome, cpf, birthDate, balance");
-            break;
-        }
-
-    };
-    case "balance":{
-        
-        const name: string = process.argv[3];
-        const cpf: string = process.argv[4];
-
-        const findCPF = accountsFile.find((account)=>{
-            return account.cpf === cpf
-        })
-
-        if(findCPF!==undefined && findCPF.name===name){
-            console.log(`O seu saldo é ${findCPF.balance}`);
-            break;
-        }else{
-            console.log(errorMensage);
-            break;
-        }
-
+        createAccount(mockAccount);
+        break;
     };
 
-    case "addMoney":{
+    case "addBalance":{
         const name: string = process.argv[3];
         const cpf: string = process.argv[4];
         const money: number = Number(process.argv[5]);
 
-        const findCPF = accountsFile.find((account)=>{
-            return account.cpf === cpf
-        })
-
-        if(findCPF!==undefined && findCPF.name===name){
-            
-            const neWFile = accountsFile.map(element=>{
-                if(findCPF===element){
-
-                    const newStatement: statement={
-                        operation: action,
-                        data: money
-                    }
-        
-                    element.balance+=money;
-                    element.statement.push(newStatement);
-                    return element;
-                }else{
-                    return element;
-                }
-            })
-
-            fs.writeFileSync(`data/accounts.json`, JSON.stringify(neWFile));
-            break;
-        }else{
-            console.log(errorMensage)
-        }
+        addBalance(name, cpf, money);
+        break;
     };
 
     case "payBill":{
+
 
     };
 
@@ -125,70 +42,12 @@ switch(action){
         const nameDestiny: string = process.argv[6];
         const cpfDestiny: string = process.argv[7];
 
-        const findCPFOrigin = accountsFile.find((account)=>{
-            return account.cpf === cpfOrigin
-        });
-
-        const findCPFDestiny = accountsFile.find((account)=>{
-            return account.cpf === cpfDestiny
-        });
-
-        if(findCPFOrigin!==undefined && findCPFDestiny!==undefined){
-            
-            if(findCPFOrigin.name === nameOrigin && findCPFDestiny.name === nameDestiny){
-
-                if(findCPFOrigin.balance>=money){
-
-                    const neWFile = accountsFile.map(element=>{
-                        if(findCPFOrigin===element){
-        
-                            const newStatement: statement={
-                                operation: action,
-                                data: money
-                            }
-                
-                            element.balance-=money;
-                            element.statement.push(newStatement);
-                            return element;
-                            
-                        }else if(findCPFDestiny===element){
-    
-                            const newStatement: statement={
-                                operation: action,
-                                data: money
-                            }
-                
-                            element.balance+=money;
-                            element.statement.push(newStatement);
-                            return element;
-                        }else{
-                            return element;
-                        }
-                    });
-                    
-                    fs.writeFileSync(`data/accounts.json`, JSON.stringify(neWFile));
-                    console.log("Transferencia realizada com sucesso")
-                    break;
-    
-                }else{
-                    console.log("Saldo insuficiente");
-                    break;
-                }
-            }else{
-                console.log(errorMensage);
-                break;
-            }
-        }else{
-            console.log(errorMensage);
-            break;
-        }
+        transference(nameOrigin, cpfOrigin, money, nameDestiny, cpfDestiny);
+        break;
     };
 
     case "getAllAccounts":{
-        const allAccounts = accountsFile.map(element=>{
-            return (`${element.name}, ${element.cpf}`)
-        });
-        console.log(allAccounts);
+        getBalance();
         break;
     }
