+++ semana14/sistema-bancario/src/services/getBalance.ts
@@ -1,0 +1,9 @@
+const accountsFile = require("../../data/accounts.json");
+
+export default function getBalance(){
+    const allAccounts = accountsFile.map(element=>{
+        return (`${element.name}, ${element.cpf}`)
+    });
+   
+    return console.log(allAccounts);   
+}
