--- semana14/sistema-bancario/src/testeInput.ts
@@ -1,56 +1,0 @@
-import * as readline from 'readline';
-
-// let rl = readline.createInterface({
-//   input: process.stdin,
-//   output: process.stdout
-// });
-
-// rl.question('Is this example useful? [y/n] ', (answer) => {
-//   switch(answer.toLowerCase()) {
-//     case 'y':
-
-//       // let rl = readline.createInterface({
-//       //   input: process.stdin,
-//       //   output: process.stdout
-//       // })
-      
-//       // rl.question("teste",(response)=>{
-//       //   switch(response){
-//       //     case "n":
-//       //       console.log("deu certo");
-//       //       break;
-//       //   }
-//       // rl.close();
-//       // })
-
-       
-//       console.log('Super!');
-//       break;
-//     case 'n':
-//       console.log('Sorry! :(');
-//       break;
-//     default:
-//       console.log('Invalid answer!');
-//   }
-//   rl.close();
-// });
-
-// function teste (){
-//   var stdin = process.openStdin();
-  
-//  return stdin.addListener("data", function(d) {
-//       // note:  d is an object, and when converted to a string it will
-//       // end with a linefeed.  so we (rather crudely) account for that  
-//       // with toString() and then substring() 
-//       console.log("you entered: [" + d.toString().trim() + "]");
-  
-      
-//   }
-//   );
-
-  
-
-// }
-
-// teste();
-
