+++ semana14/sistema-bancario/src/types.ts
@@ -1,0 +1,12 @@
+export type account = {
+    name: string,
+    cpf: string,
+    birthDate: string,
+    balance: number,
+    statement: statement[]
+}
+
+export type statement = {
+    operation: string,
+    data: any
+}
